requirements.txt
streamlit==1.42.1
google-genai
pydantic
pandas
plotly
PyPDF2
python-dotenv
bcrypt
